Change popup items from "<" to "less than", etc.
On some operating systems (Mac OS) and some browser versions (Safari version??), the characters "<", "=", ">" are not being sent to the server properly so the answer is not submitted properly to the server (the answer submitted shows up as blank).  This was resolved by replacing the special characters "<", "=" and ">" by "less than", "equal to" and "greater than".