Correct and (partially) polish problem.
Clean up typesetting and insure that the
leading 2 x 2 leading principal minor of
the coefficient matrix does not vanish.
(Otherwise the question has infinitely
many correct answers, only one of which
is accepted.)  Remove the MathObjects
comment, since this problem does not
use MathObjects.

When the commits in this branch have
been applied, Bug 2052 should be closed.