Smaller domain overall, abs tol 0.001 in part c
The correct answer in part c does uses decimal approximations for some of the coefficients instead of fractions that are accurate to infinitely many decimal places.  If a student enters the correct answer using fraction coefficients, their answer may be marked incorrect.  The smaller domain and the tighter tolerances are an attempt to make the answer checker work properly all the time.  In an ideal world, I would rewrite this problem from scratch with the answer checking in mind so as to avoid problems like this.  But, changing the domain and tolerances should work well enough.  (We'll see how many people report problems.)