Change so that answers are order 10^7 not 10^26
The problem previously had answers of the order 10^26 with margin of error of the order 0.1 percent of 10^26 (in other words order 10^23).  Yikes!  By decreasing exponents from 10 and 12 down to 3 and 2, the problem now has answers of order 10^7, and switching to absolute tolerance the margin of error is 0.001 (for an answer that's a large integer).  Previously, the correct answer was not being marked correct (because $x1 can be negative and the original problem author did not put parentheses around it, causing errors such as -4^2 instead of (-4)^2 appearing as factors).  Also, I updated to use MathObjects more extensively.  All these changes necessitated removing the "report to 3 significant digits" instructions from the problem text, as the answer is now something students should be able to find exactly (it's a large integer) and enter without using scientific notation.