remove quotes around scalar $answertolerance
Single quotes: '$answertolerance' is the literal string $answertolerance, so it does not make sense for the answer checker.
Double quotes: "$answertolerance" substitutes the numerical value of $answertolerance and then the quotes convert the result to a Perl string (an extra, unnecessary step).
No quotes: $answertolerance substitutes the numerical value, as desired.

The best solution is  tolerance => $answertolerance with no quotes.  I have checked it and it works.

Good catch on the single quotes, George!