Change self-referential naming
The function to differentiate was named g(x).  The quotient rule formula provided uses [f/g]', which may confuse students since the function given is named g(x).  So, I changed the name of the function to be differentiated to h(x), so that students can interpret it as h(x) = f(x) / g(x) instead of g(x) = f(x) / g(x) when they apply the quotient rule.