Restrict domain of function evaluation
With seed 2875 the correct answer for the first part was not being marked correct.  Because the answer had e^(-7t) and e^(-9t) in it, I suspect that for values of t in [0,2] the answer was below the zero level tolerance.  I restricted the domain to t in [0,1], which makes the answer checker mark the correct answer as correct for this seed.