Adjust answer checker
Seed 4557 has a particular solution that is constant.  For this seed, the answer checker was marking a constant particular solution incorrect.  I reconfigured the answer checker (moving the vector g to the other side of the equation) to avoid zero-level tolerance errors when checking the particular solution.  I also removed the display of the "Entered Answer" because it was formatted poorly and because  the "Preview Answer" was so lengthy.