Fix bug in logarithmic differentiation problem.
One of my students got a "Can't generate enough valid points for comparison"
error when inputting the correct answer for this problem.  They used
e^(7ln(x)/x) instead of the equivalent x^(7/x) in their answer, so perhaps
that was what caused the problem.

The error disappears when we change the limits to [1,5] to ensure positive
inputs to ln.