Add missing image to related rates problem.
This related rates problem is missing an image which is essential for
completing the problem, as I discovered while working with a student
today in office hours!  Without it, there is not enough information relating
the volume and height of the water.

I referenced the corresponding problem in my hardcopy of Stewart to create
this dynamic image instead of uploading a scanned image.  This way, if this
problem is ever updated to use randomly-generated numbers instead of hardcoded
ones, the image can be updated automatically.