Add several problems where students compute "at least one" probabilities.
We use Stevens' Introduction to Statistics: Think & Do for our introductory
statistics course.  It has an entire section devoted to computing
probabilities of an event occuring at least once, i.e., realizing that this
is the complement of the event never occuring.

There are few problems of this sort currently in the OPL, so I adapted some
from Stevens.