Detect equivalent solutions to implicit differentiation problem.
Students will get different, but equivalent, answers to this problem
depending on whether they first multiply everything by the denominator of
the left-hand side of the equation before differentiating.

Previously, a student's answer would only be marked correct if they did
in fact do this.  If a student did not do this, and instead immediately
began differentiating using the quotient rule, then their answer would be
marked incorrect even if they found the derivative correctly.

While seemingly very different, the two answers obtained by these two
methods are equivalent if one substitutes on side of the original implicit
equation for the other.

However, previously the two test points for x and y are chosen at random,
and not with the implicit relationship from the original equation in mind.

We fix this by making sure that the test points for x and y which are chosen
satisfy the original equation.  The code is based on Alex Jordan's
excellent answer to my question on the WeBWorK forums [1].

[1] http://webwork.maa.org/moodle/mod/forum/discuss.php?d=4574