Switch to absolute tolerance in part b
With seed 4882, one of the random x-values is x=0.2.  When the correct answer (6-30x)e^(5x) is evaluated at x=0.2 the result is -1.44859E-14.  When the algebraically equivalent answer 6e^(5x)-30xe^(5x) entered by a student is evaluated at x=0.2 the result is -1.42109E-14.  In this case, the absolute error 2.75061E-16 is acceptable, but the relative error 0.0189882... is not acceptable, and the student's answer is marked incorrect.  I speculate that this is due to numerical instability and internal rounding errors (in decimal to binary conversion).  In any case, a good fix is to use absolute tolerance for part b with a reasonable tolerance level.