Bug fix
This problem was using an undefined variable $Earth_radius. grep-ping, I find that used in one other problem in this collection, and in that other problem it is defined as 6.38. But based on this problem's answer, I think the intention is to use 6.4. Just changing the variable reference to simply use 6.4.