Fix bug in playing card probability problem.
The goal of the problem is to have a hand of four consecutive cards in
the same suit, none of which are aces to allow for the possibility
of a straight with either an ace low or an ace high.

However, there was a chance that the hand could have begun with a
a jack, which meant that our highest card was already an ace and
so our probability of a straight is reduced by half.  There was
also a possibility that our hand began with a queen or king, which
ended up displaying empty strings for the highest card(s) in our
hand.

This has been fixed by ensuring that the highest card picked for
starting the hand is a 10.  Also, 'Ace' has been removed from the
list of ranks as it will never need to be displayed.