Possible incorrect conditional probability
I always find these probability problems to be quite tricky.  Hopefully this analysis is correct.

But I believe that since this is worded as the conditional probability P(rej|pos), then if this is supposed to be a Bayes' theorem problem we would have to turn it into something involving P(pos) in a law of total probability sense.  That is, P(rej|pos) = P(pos|rej)*P(rej) /  P(pos), where P(pos) = P(pos|rej)*P(rej) + P(pos|not-rej)*P(not-rej).  In words, the probability of a positive test is the sum of the situation where it's positive and actually being rejected and the one where it's positive and not being actually rejected.

The probability of P(pos|not-rej) is indeed `$d`, and `$e` and `$g` are defined to be the complementary probabilities of rejection and non-rejection.  But P(pos|rej) should be `1-$d`, since both of them are about positive tests.  The negative test number hardcoded 20 shouldn't be relevant.

My apologies for putting this all in a commit message, but I didn't want to clutter up GH by putting in a separate issue and PR when I had found the solution, or so I believe.