Really fix bug 3898 by making all the numbers smaller.  Seed 1950 for example was still broken because the answer was on the order of 10^10 (and 10^20 was possible with some seeds) and sometimes the problem logic was not taking it.  I have 11 students, one had issues.  Entering the exact answer for seed 1950 was not being accepted.
So make all numbers smaller.  The largest possible answer is now
500000, which does not cause trouble (tested).  Actually I've tested
with a bunch of seeds just to be sure.

There is really no reason to have more possible answers than atoms in
the universe.  It's less now, but there are still a huge number of
possible problem variations (hundereds of thousands at least, if I am
counting right).

While there, I also made sure none of the coordinates are 0, so that
the function is not trivially zero at one or the other point, making
sure that the student understands that it must be evaluated at both
points.

And also while there, I improved the spacing somewhat and fixed a
commented out TEXT(beginproblem());