Update SCalcET7-11-1-ex3.pg
The original problem design assumes that the reciprocal of an increasing function is necessarily decreasing.  That is not so for functions which are neither strictly positive nor strictly negative. The proposed fix sets the variable $c to be no longer random, in order to make sure that the denominator of each term of the sequence is strictly positive.