Update APEX_3.3_20.pg
Not sure if this is a bug or a matter of different conventions adopted by different institutions.  If a function is increasing on (a, b) and on (c, d), can we say in general that it is increasing on (a, b) U (c, d)? I would say no, but others might disagree?