Update 4.4.9.pg
On line 57, it is possible to be so unlucky that the output of non_zero_random(0, $cp->eval()) is exactly $cp->eval(). In that case, the problem could think there is no inflection point when there actually is.

I am not sure if it is safe to pass a non-integer argument to the non_zero_random() function. Someone more knowledgeable should probably double-check this...