There are nomenclature differences about which matrix is the the "change of basis matrix from B to C". The disagreement in nomenclature makes it dangerous to request the change of basis matrice by "name". Insead, request them by the "property" of how the transform coordinate vectors.
In addition, modify the parameter generation code to (a) prevent the
two COB matrices from being the same for any student, and (b) limit the
number of loop iterations and use a fixed fallback when needed.

The nomenclature issue was raised by Vadim Panfilov vadim.panfilov@gtiit.edu.cn
in a correspondence, as his course used the "other" nomenclature.

The two different choices of nomenclature are discussed at
https://math.stackexchange.com/questions/59339/change-of-basis-confusion
and
https://math.stackexchange.com/questions/21752/why-is-the-change-of-basis-matrix-called-such
and different text-books / authors take different sides of the matter.