Fix statistics problem so that data set will only have one mode
This problem constructs an array @<!-- -->data so that $data[5],
$data[6], and $data[7] will always hold the same value, and therefore
this value will appear in the data set three times and ideally end up
being the mode of the given data set.

However, the gaps between $data[15] and $data[16] and between
$data[16] and $data[17] will both be 0 with probability 1/16, giving
us a second mode, which makes the "correct" answer of $data[5]
invalid.

We change the final gap between $data[16] and $data[17] so that it is
always positive, removing this possibility.