Update Basis_30.pg
The original "correct" answer "<1, 0, 0>, <0, 1, 0>" implies that the 3-rd entry of every vector in the column space is equal to 0. We cannot know that based on the RREF of the matrix.

Also, cast a1, a2, ... as "Vector3D", otherwise an (incorrect) answer of the form <a, b, c> would throw an error message indicating wrong answer type.