Fixed Answer Checker for $ans4 in Part (c).
I earlier mistakenly removed the answer checker for $ans4 in Part (c), having overlooked the wording of the question. (The answer can be a non-unique combination of a1, a2, and a3.)
The answer checker has been put back in, and modified to work with a1, a2,... now being  "Vector3D" (instead of 'Real') variables.

Also defined $a1, $a2,... $zero as global PG variables, since they are now used in two different answer checkers. 

For $ans4 in Part (c), if an incorrect answer was entered first, and the a correct answer is entered, a message "This answer is equivalent to the one you just submitted." appears, though both attempts are marked "incorrect" and "correct" respectively, which is expected behaviour. I don't know what's causing this.