Clarify depth in fluid force problem
With the original wording, it's not clear exactly how the plate is
submerged.  Is the top of the plate, the bottom, or maybe even the
middle of the plate at the given depth?

Based on the solution, it is the bottom of the plate that's submerged
to this depth, so we clarify that in the wording of the problem.