Prevent point placed with label outside of window
The problem generates two random but distinct numbers between 2 and 5.
In order to assure that they are distinct, it increases one of them by
1 if they are equal. This may result in one of the numbers being 6,
which will cause a point on the graph placed on the edge of the graphing
window, with invisible label, which makes it impossible to answer the
problem.

The fix selects the second number repeatedly until the numbers are
distinct.